+++ tests/testthat/test-ipmr_generics.R
@@ -214,8 +214,8 @@ test_that("getters and setters work as expected", {
   test_exprs <- exprs(
     s = exp(s_i + s_s * size_1) / (1 + exp(s_i + s_s * size_1)),
     mu_g = g_int + g_slope * size_1,
-    f_d_1 = stats::dunif(size_2, 0.15, 0.25),
-    f_d_2 = stats::dnorm(size_2, mu_f_d_2, sd_f_d_2)
+    f_d_1 = dunif(size_2, 0.15, 0.25),
+    f_d_2 = dnorm(size_2, mu_f_d_2, sd_f_d_2)
   )
 
   easterling_test <- pdb_exprs$aaa310[c("s", "mu_g", "f_d_1", "f_d_2")]
@@ -345,7 +345,8 @@ test_that("getters and setters work as expected", {
   expect_s3_class(pdb_pars$aaaa17, "ipmr_parameters")
   expect_s3_class(pdb_pars$aaa310, "ipmr_parameters")
 
-  expect_equal(pdb_pars$aaa310$s_i, 1.34, ignore_attr = "flat_protect")
+  expect_equal(pdb_pars$aaa310$s_i, 1.34,
+               ignore_attr = c("flat_protect", "na_ok"))
 
   parameters(y) <- list(aaa310 = list(s_i = 1.6),
                         aaaa16 = list(g_i = 12,
@@ -353,12 +354,15 @@ test_that("getters and setters work as expected", {
 
   new_pars <- parameters(y)$aaa310
 
-  expect_equal(new_pars$s_i, 1.6, ignore_attr = "flat_protect")
+  expect_equal(new_pars$s_i, 1.6,
+               ignore_attr = c("flat_protect", "na_ok"))
 
   new_pars <- parameters(y)$aaaa16
 
-  expect_equal(new_pars$g_i, 12, ignore_attr = "flat_protect")
-  expect_equal(new_pars$g_s, 14, ignore_attr = "flat_protect")
+  expect_equal(new_pars$g_i, 12,
+               ignore_attr = c("flat_protect", "na_ok"))
+  expect_equal(new_pars$g_s, 14,
+               ignore_attr = c("flat_protect", "na_ok"))
 
   # pop_state