+++ NEWS.md
@@ -1,5 +1,27 @@
+# Rpadrino 0.0.4
+
+This contains new features, and a couple changes for compatibility with `ipmr 0.0.5`. 
+
+## Features
+
+- Accessor functions for the Metadata table now have `ipm_id`s as names. This enables patterns for subsetting like this:
+
+```
+spps <- pdb_species_accepted(pdb)
+ids  <- names(spps)[spps %in% c("Carpobrotus_spp", "Geum_radiatum")]
+my_proto_ipms <- pdb_make_proto_ipm(pdb, ipm_id = ids)
+
+```
+
+- New vignettes on working with stochastic IPMs, and on the messages that are produced while building IPMs with `pdb_make_proto_ipm()` and `pdb_make_ipm()`.
+
+- `pdb_report()` now contains equations for each IPM using `ipmr`'s `make_ipm_report_body()`. 
+
+    + NB: This is currently unavailable for [parameter set indexed IPMs](https://levisc8.github.io/ipmr/articles/index-notation.html).
+
 # Rpadrino 0.0.3
 
+- Fixed some issues from two of CRAN's Linux machines.
 
 # Rpadrino 0.0.2